"Add InputField component
Created a new file 'InputField.js' in 'src/components/InputField'. This new React component will render an input field for entering a company name. This change was motivated by the need to provide a user interface for inputting company names. This is a new addition to the codebase and does not modify any existing functionality."