"Add ResultsTable component"
This commit introduces a new React component, ResultsTable, located in src/components/ResultsTable/ResultsTable.js. The motivation for this change was to create a component that can render a responsive table for displaying results. This is a new addition to the codebase and does not modify any existing implementation. The ResultsTable component will be used to display data in an easy-to-read, clean layout, adapting to various screen sizes.