"Implement express server and POST endpoint in server.js"
This commit introduces a new express server setup and a POST endpoint '/api/leadership' in the server.js file. The endpoint takes a company name as input and uses the OpenAI module to search for leadership information about the company. The server listens on port 3000 by default.

The motivation for this change was to provide a backend service for our React application that can handle API requests and return leadership information about a company based on the company name provided. This is a new implementation, as the server.js file was just created and did not have any previous functionality.

This change is part of a larger task to create a React application that allows users to enter a company name and receive leadership information about the company. Other components of the application, such as the 'App', 'InputField', and 'ResultsTable' components, have also been created or modified as part of this task.