"Change fetch API request method to POST in App.js
The motivation for this change was to allow the sending of data (companyName) in the request body. Previously, the fetch API request was using the 'GET' method, which does not support sending data in the body. Now, the method has been changed to 'POST' to accommodate this requirement. This change is part of a series of updates to the App component to handle user input and API requests for company leadership information."