"Add SearchCompany.js component
This commit includes the creation of a new React component, SearchCompany.js, which includes an input field for users to enter a company name. This change was motivated by the need to provide a user interface for searching companies. This is a new addition to the codebase and does not modify any existing implementation."